Move utils package to internal
